+++ src/senzing/G2ConfigMgr.py
@@ -1,11 +1,10 @@
 from ctypes import *
 import threading
-import json
 import os
 import functools
 import warnings
 
-from .G2Exception import TranslateG2ModuleException, G2ModuleNotInitialized, G2ModuleGenericException
+from .G2Exception import TranslateG2ModuleException, G2NotInitializedException, G2Exception
 
 __all__ = ['G2ConfigMgr']
 SENZING_PRODUCT_ID = "5041"  # See https://github.com/Senzing/knowledge-base/blob/main/lists/senzing-component-ids.md
@@ -95,7 +94,7 @@ class G2ConfigMgr(object):
             print("ERROR: Unable to load G2.  Did you remember to setup your environment by sourcing the setupEnv file?")
             print("ERROR: For more information see https://senzing.zendesk.com/hc/en-us/articles/115002408867-Introduction-G2-Quickstart")
             print("ERROR: If you are running Ubuntu or Debian please also review the ssl and crypto information at https://senzing.zendesk.com/hc/en-us/articles/115010259947-System-Requirements")
-            raise G2ModuleGenericException("Failed to load the G2 library")
+            raise G2Exception("Failed to load the G2 library")
 
         self._resize_func_def = CFUNCTYPE(c_char_p, c_char_p, c_size_t)
         self._resize_func = self._resize_func_def(resize_return_buffer)
@@ -185,7 +184,7 @@ class G2ConfigMgr(object):
         ret_code = self._lib_handle.G2ConfigMgr_addConfig(_configStr, _configComments, cID)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2ConfigMgr has not been successfully initialized')
+            raise G2NotInitializedException('G2ConfigMgr has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2ConfigMgr_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -204,7 +203,7 @@ class G2ConfigMgr(object):
         ret_code = self._lib_handle.G2ConfigMgr_getConfig(configID_, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2ConfigMgr has not been successfully initialized')
+            raise G2NotInitializedException('G2ConfigMgr has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2ConfigMgr_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -223,7 +222,7 @@ class G2ConfigMgr(object):
         ret_code = self._lib_handle.G2ConfigMgr_getConfigList(pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2ConfigMgr has not been successfully initialized')
+            raise G2NotInitializedException('G2ConfigMgr has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2ConfigMgr_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -240,7 +239,7 @@ class G2ConfigMgr(object):
         ret_code = self._lib_handle.G2ConfigMgr_setDefaultConfigID(configID_)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2ConfigMgr has not been successfully initialized')
+            raise G2NotInitializedException('G2ConfigMgr has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2ConfigMgr_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -255,7 +254,7 @@ class G2ConfigMgr(object):
         ret_code = self._lib_handle.G2ConfigMgr_replaceDefaultConfigID(oldConfigID_, newConfigID_)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2ConfigMgr has not been successfully initialized')
+            raise G2NotInitializedException('G2ConfigMgr has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2ConfigMgr_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -270,7 +269,7 @@ class G2ConfigMgr(object):
         ret_code = self._lib_handle.G2ConfigMgr_getDefaultConfigID(cID)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2ConfigMgr has not been successfully initialized')
+            raise G2NotInitializedException('G2ConfigMgr has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2ConfigMgr_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)