+++ src/senzing/G2Engine.py
@@ -1,13 +1,11 @@
 from ctypes import *
 import threading
-import json
 import os
 import functools
 import warnings
 
-from csv import reader as csvreader
 
-from .G2Exception import TranslateG2ModuleException, G2ModuleNotInitialized, G2ModuleGenericException
+from .G2Exception import TranslateG2ModuleException, G2NotInitializedException, G2Exception
 from .G2EngineFlags import G2EngineFlags
 
 __all__ = ['G2Engine']
@@ -131,7 +129,7 @@ class G2Engine(object):
             print("ERROR: Unable to load G2.  Did you remember to setup your environment by sourcing the setupEnv file?")
             print("ERROR: For more information see https://senzing.zendesk.com/hc/en-us/articles/115002408867-Introduction-G2-Quickstart")
             print("ERROR: If you are running Ubuntu or Debian please also review the ssl and crypto information at https://senzing.zendesk.com/hc/en-us/articles/115010259947-System-Requirements")
-            raise G2ModuleGenericException("Failed to load the G2 library")
+            raise G2Exception("Failed to load the G2 library")
 
         self._resize_func_def = CFUNCTYPE(c_char_p, c_char_p, c_size_t)
         self._resize_func = self._resize_func_def(resize_return_buffer)
@@ -285,7 +283,7 @@ class G2Engine(object):
             print("Initialization Status: " + str(ret_code))
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -308,7 +306,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_process(input_umf_string)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -335,7 +333,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_processWithInfo(input_umf_string, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -369,7 +367,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_processWithResponseResize(input_umf_string, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -393,7 +391,7 @@ class G2Engine(object):
         self._lib_handle.G2_checkRecord.argtypes = [c_char_p, c_char_p, POINTER(c_char_p), POINTER(c_size_t), self._resize_func_def]
         ret_code = self._lib_handle.G2_checkRecord(_inputUmfString, _recordQueryList, pointer(responseBuf), pointer(responseSize), self._resize_func)
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -413,7 +411,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_exportJSONEntityReport(flags, byref(exportHandle))
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -436,7 +434,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_exportCSVEntityReport(_headersForCSV, flags, byref(exportHandle))
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -458,7 +456,7 @@ class G2Engine(object):
         while resultValue != 0:
 
             if resultValue == -1:
-                raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+                raise G2NotInitializedException('G2Engine has not been successfully initialized')
             elif resultValue < 0:
                 self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
                 raise TranslateG2ModuleException(tls_var.buf.value)
@@ -495,7 +493,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_addRecord(_dataSourceCode, _recordId, _jsonData, _load_id)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -519,7 +517,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_addRecordWithReturnedRecordID(_dataSourceCode, _jsonData, _load_id, tls_var.buf, sizeof(tls_var.buf))
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -550,7 +548,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_addRecordWithInfo(_dataSourceCode, _recordId, _jsonData, _load_id, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -588,7 +586,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_addRecordWithInfoWithReturnedRecordID(_dataSourceCode, _jsonData, _load_id, flags, tls_var.buf, sizeof(tls_var.buf), pointer(infoBuf), pointer(infoBufSize), self._resize_func3)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -614,7 +612,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_replaceRecord(_dataSourceCode, _recordId, _jsonData, _load_id)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -643,7 +641,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_replaceRecordWithInfo(_dataSourceCode, _recordId, _jsonData, _load_id, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -666,7 +664,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_deleteRecord(_dataSourceCode, _recordId, _load_id)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -692,7 +690,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_deleteRecordWithInfo(_dataSourceCode, _recordId, _load_id, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -716,7 +714,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_reevaluateRecord(_dataSourceCode, _recordId, flags)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -740,7 +738,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_reevaluateRecordWithInfo(_dataSourceCode, _recordId, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -760,7 +758,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_reevaluateEntity(entityID, flags)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -782,7 +780,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_reevaluateEntityWithInfo(entityID, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -810,7 +808,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_searchByAttributes_V2(_jsonData, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -835,7 +833,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_searchByAttributes_V3(_jsonData, _searchProfile, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -864,7 +862,7 @@ class G2Engine(object):
         self._lib_handle.G2_findPathByEntityID_V2.argtypes = [c_longlong, c_longlong, c_int, c_longlong, POINTER(c_char_p), POINTER(c_size_t), self._resize_func_def]
         ret_code = self._lib_handle.G2_findPathByEntityID_V2(startEntityID, endEntityID, maxDegree, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -895,7 +893,7 @@ class G2Engine(object):
         self._lib_handle.G2_findNetworkByEntityID_V2.argtypes = [c_char_p, c_int, c_int, c_int, c_longlong, POINTER(c_char_p), POINTER(c_size_t), self._resize_func_def]
         ret_code = self._lib_handle.G2_findNetworkByEntityID_V2(_entityList, maxDegree, buildOutDegree, maxEntities, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -930,7 +928,7 @@ class G2Engine(object):
         self._lib_handle.G2_findPathByRecordID_V2.argtypes = [c_char_p, c_char_p, c_char_p, c_char_p, c_int, c_longlong, POINTER(c_char_p), POINTER(c_size_t), self._resize_func_def]
         ret_code = self._lib_handle.G2_findPathByRecordID_V2(_startDsrcCode, _startRecordId, _endDsrcCode, _endRecordId, maxDegree, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -961,7 +959,7 @@ class G2Engine(object):
         self._lib_handle.G2_findNetworkByRecordID_V2.argtypes = [c_char_p, c_int, c_int, c_int, c_longlong, POINTER(c_char_p), POINTER(c_size_t), self._resize_func_def]
         ret_code = self._lib_handle.G2_findNetworkByRecordID_V2(_recordList, maxDegree, buildOutDegree, maxEntities, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -980,7 +978,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_whyRecordInEntity_V2(_dataSourceCode, _recordID, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1003,7 +1001,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_whyEntityByRecordID_V2(_dataSourceCode, _recordID, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1023,7 +1021,7 @@ class G2Engine(object):
         self._lib_handle.G2_whyEntityByEntityID_V2.argtypes = [c_longlong, c_longlong, POINTER(c_char_p), POINTER(c_size_t), self._resize_func_def]
         ret_code = self._lib_handle.G2_whyEntityByEntityID_V2(entityID, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1043,7 +1041,7 @@ class G2Engine(object):
         self._lib_handle.G2_howEntityByEntityID_V2.argtypes = [c_longlong, c_longlong, POINTER(c_char_p), POINTER(c_size_t), self._resize_func_def]
         ret_code = self._lib_handle.G2_howEntityByEntityID_V2(entityID, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1065,7 +1063,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_getVirtualEntityByRecordID_V2(_recordList, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1086,7 +1084,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_whyEntities_V2(entityID1, entityID2, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1111,7 +1109,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_whyRecords_V2(_dataSourceCode1, _recordID1, _dataSourceCode2, _recordID2, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1143,7 +1141,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_findPathExcludingByEntityID_V2(startEntityID, endEntityID, maxDegree, _excludedEntities, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1177,7 +1175,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_findPathIncludingSourceByEntityID_V2(startEntityID, endEntityID, maxDegree, _excludedEntities, _requiredDsrcs, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1215,7 +1213,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_findPathExcludingByRecordID_V2(_startDsrcCode, _startRecordId, _endDsrcCode, _endRecordId, maxDegree, _excludedEntities, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1255,7 +1253,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_findPathIncludingSourceByRecordID_V2(_startDsrcCode, _startRecordId, _endDsrcCode, _endRecordId, maxDegree, _excludedEntities, _requiredDsrcs, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1284,7 +1282,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_getEntityByEntityID_V2(entityID, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1316,7 +1314,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_getEntityByRecordID_V2(_dsrcCode, _recordId, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1340,7 +1338,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_findInterestingEntitiesByEntityID(entityID, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1368,7 +1366,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_findInterestingEntitiesByRecordID(_dsrcCode, _recordId, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1390,7 +1388,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_getRedoRecord(pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1412,7 +1410,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_processRedoRecord(pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1460,7 +1458,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_countRedoRecords()
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1491,7 +1489,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_getRecord_V2(_dsrcCode, _recordId, flags, pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1513,7 +1511,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_stats(pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1535,7 +1533,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_exportConfig(pointer(responseBuf), pointer(responseSize), self._resize_func)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1549,7 +1547,7 @@ class G2Engine(object):
             ret_code2 = self._lib_handle.G2_getActiveConfigID(cID)
 
             if ret_code2 == -1:
-                raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+                raise G2NotInitializedException('G2Engine has not been successfully initialized')
             elif ret_code2 < 0:
                 self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
                 raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1570,7 +1568,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_getActiveConfigID(cID)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1591,7 +1589,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_getRepositoryLastModifiedTime(lastModifiedTimeStamp)
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)
@@ -1612,7 +1610,7 @@ class G2Engine(object):
         ret_code = self._lib_handle.G2_purgeRepository()
 
         if ret_code == -1:
-            raise G2ModuleNotInitialized('G2Engine has not been successfully initialized')
+            raise G2NotInitializedException('G2Engine has not been successfully initialized')
         elif ret_code < 0:
             self._lib_handle.G2_getLastException(tls_var.buf, sizeof(tls_var.buf))
             raise TranslateG2ModuleException(tls_var.buf.value)